# docs/swagger.yaml
basePath: /api/v1
definitions:
  entity.Channel:
    properties:
      body:
        type: string
      subject:
        type: string
    type: object
  entity.Channels:
    properties:
      email:
        $ref: '#/definitions/entity.Channel'
      phone:
        $ref: '#/definitions/entity.Channel'
    type: object
  entity.EmailPreferences:
    properties:
      approval:
        type: boolean
      notifyType:
        type: string
    type: object
  entity.PhonePreferences:
    properties:
      approval:
        type: boolean
      notifyType:
        type: string
    type: object
  entity.Preferences:
    properties:
      email:
        $ref: '#/definitions/entity.EmailPreferences'
      phone:
        $ref: '#/definitions/entity.PhonePreferences'
    type: object
  entity.RequestNotification:
    properties:
      channels:
        $ref: '#/definitions/entity.Channels'
      notifyType:
        type: string
      userId:
        type: string
    type: object
  entity.RequestPreferences:
    properties:
      preferences:
        $ref: '#/definitions/entity.Preferences'
      userId:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Notification Service
  title: Notify Service
  version: v1.0.0
paths:
  /messages/send:
    post:
      consumes:
      - application/json
      description: Sends a message to the user with the specified fields.
      parameters:
      - description: send mess
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.RequestNotification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
        "500":
          description: Failed to create notification
        "504":
          description: Waiting time is up
      summary: Sending a message
      tags:
      - Notify
  /user/preferences:
    put:
      consumes:
      - application/json
      description: Changes user preferences
      parameters:
      - description: user preferences
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.RequestPreferences'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
        "500":
          description: Failed to edit preferences
        "504":
          description: Waiting time is up
      summary: Changing preferences
      tags:
      - Preferences
swagger: "2.0"